requirements.txt
django-celery==3.1
